make --schema arg multiOption
