+++ lib/src/schema_converter.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:path/path.dart' as p;
 import 'package:schema_dart/src/logger.dart';
-import 'package:schema_dart/src/table_reader.dart';
+import 'package:schema_dart/src/postgres_reader.dart';
 import 'package:schema_dart/src/types_generator.dart';
 
 import 'types.dart';
@@ -11,31 +11,42 @@ class SchemaConverter {
   final String connectionString;
   final Directory outputDirectory;
 
-  /// the schema to generate data classes from
-  final String schemaName;
+  late final PostgresReader postgresReader;
+
+  /// an optional list of schemas to generate data classes from
+  final List<String>? schemaNames;
+  final _schemas = <String>[];
 
   /// an optional list of tables to generate the data classes from
   final List<String>? tableNames;
-
   final _tables = <Table>[];
 
   SchemaConverter({
     required this.connectionString,
     required this.outputDirectory,
-    required this.schemaName,
+    this.schemaNames,
     this.tableNames,
   });
 
   Future<void> convert() async {
+    // read schemas
+    Log.trace('started reading schemas');
+    var progress = Log.progress('reading schemas');
+    await _readSchemas();
+    progress.finish(message: 'found ${-_schemas.length} - schemas');
+
     // read tables
     Log.trace('started reading tables');
-    var progress = Log.progress('reading tables');
+    progress = Log.progress('reading tables');
     await _readTables();
-    progress.finish(message: 'found ${_tables.length} - tables');
+    progress.finish(message: 'found ${_tables.length} - tables in $_schemas');
+
+    Log.trace('disconnecting the database');
+    await postgresReader.disconnect();
 
     // generate source code
     Log.trace('generating dart classes');
-    progress = Log.progress('generating dart source code for tables');
+    progress = Log.progress('generating dart source code ®for tables');
     await _addDartSourceToTables();
     progress.finish(message: 'sources were generated.');
 
@@ -46,14 +57,19 @@ class SchemaConverter {
     progress.finish(message: 'files were written at ${outputDirectory.path}');
   }
 
-  Future<void> _readTables() async {
-    final reader = TablesReader.fromConnectionString(connectionString);
+  Future<void> _readSchemas() async {
+    postgresReader = PostgresReader.fromConnectionString(connectionString);
     Log.trace('connecting to database');
-    await reader.connect();
+    await postgresReader.connect();
+    Log.trace('calling PostgresReader.getSchemas');
+    final schemas = await postgresReader.getSchemas(schemaNames: schemaNames);
+    _schemas.clear();
+    _schemas.addAll(schemas);
+  }
+
+  Future<void> _readTables() async {
     Log.trace('calling TablesReader.getTables');
-    final tables = await reader.getTables(schemaName: schemaName, tableNames: tableNames);
-    Log.trace('disconnecting the database');
-    await reader.disconnect();
+    final tables = await postgresReader.getTables(schemaNames: _schemas, tableNames: tableNames);
     _tables.clear();
     _tables.addAll(tables);
   }
