+++ lib/src/types.dart
@@ -2,18 +2,19 @@ import 'util.dart';
 
 /// a simple representation of a table
 class Table {
+  final String schemaName;
   final String tableName;
   final List<ColumnData> columns;
 
   String source = '';
 
-  Table(this.tableName, this.columns);
+  Table(this.schemaName, this.tableName, this.columns);
 
   String get dartClassName => tableName.convertSnakeCaseToCamelCase().toUpperCaseFirst();
 
   @override
   String toString() {
-    return 'Table(tableName: $tableName, columns: ${columns.length})';
+    return 'Table(schemaName: $schemaName, tableName: $tableName, columns: ${columns.length})';
   }
 }
 
@@ -124,6 +125,7 @@ String getDartType(String postgresType, bool isNullable) {
 }
 
 class InfoSchemaColumnNames {
+  static const schemaName = 'schema_name';
   static const tableName = 'table_name';
   static const columnName = 'column_name';
   static const dataType = 'udt_name'; // do not use `data_type` (see: NOTES.md)
